refactor(admin-api): extract Servers/Ops actions; move ajax controllers to Admin\Ajax
Second slice of the admin/api.php readability refactor, plus housekeeping.

- Relocate the extracted ajax controllers into a dedicated
  XcVm\Public\Controllers\Admin\Ajax namespace/folder (BaseAjaxController,
  CacheAjaxController moved here from the previous slice).
- Rewrite the ajax-controller docblocks/comments in English (repo standard
  going forward).
- Add BaseAjaxController::gateAny([[type,key],...]) for the OR-permission
  idiom `if (check(a) || check(b))` used by the server-view family.
- Add ServerAjaxController with the 17 Servers/Ops actions: rtmp_ip,
  rollback_versions, server, proxy, fingerprint, restart_all_services,
  restart_services, reboot_server, update_binaries, server_view, server_stats,
  rtmp_kill, install_status, reinstall_server, fpm_status, update_all_servers,
  update_all_binaries. Block logic ported verbatim; the repeated server_id
  normalize idiom folded into a private helper.
- Register the 17 actions via $router->api(...) in routes/admin.php and delete
  their blocks from api.php (4825 -> 4199 lines; 786 removed across both slices).

Two deliberate departures from a byte-for-byte port, both documented in the
controller:
- rtmp_kill echoes the raw ApiClient::systemRequest() response (not JSON), as
  api.php did.
- proxy->kill previously read an undefined $rServerID (a latent bug:
  getAll()[null] -> no-op + warnings in redis-handler mode); it now uses
  server_id so the redis path actually closes the proxy's connections. This is
  a behaviour change on that path — flagged for review.

Verified: php -l, make gates (incl. check-procedural-use), make phpstan (0),
make cs. Actions still gated identically and reached via Router::dispatchApi()
before the AjaxController fallback.
