+++ src/Public/Views/admin/api.php
@@ -855,26 +855,6 @@ if (isset($_SESSION['hash'])) {
 
 			exit();
 		}
-		if (RequestManager::get('action') == 'rtmp_ip') {
-			if (Authorization::check('adv', 'add_rtmp')) {
-				$rSub = RequestManager::get('sub');
-
-				if ($rSub == 'delete') {
-					BlocklistService::deleteRTMPIP(RequestManager::get('ip'));
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				echo json_encode(array('result' => false));
-
-				exit();
-			}
-
-			echo json_encode(array('result' => false));
-
-			exit();
-		}
 		if (RequestManager::get('action') == 'watch_output') {
 			if (Authorization::check('adv', 'folder_watch_output')) {
 				$rSub = RequestManager::get('sub');
@@ -894,267 +874,6 @@ if (isset($_SESSION['hash'])) {
 			echo json_encode(array('result' => false));
 
 			exit();
-		}
-		if (RequestManager::get('action') == 'rollback_versions') {
-			if (Authorization::check('adv', 'edit_server')) {
-				// Optional base version (a specific server's version); defaults to
-				// the MAIN panel version. Previous releases are resolved relative to it.
-				$rBaseVersion = trim((string) RequestManager::get('version'));
-				if (!preg_match('/^\d+\.\d+\.\d+$/', $rBaseVersion)) {
-					$rBaseVersion = XC_VM_VERSION;
-				}
-				$rVersions = array();
-
-				try {
-					$rGit = new GitHubReleases(GIT_OWNER, GIT_REPO_MAIN, SettingsManager::get('update_channel'));
-					$rGit->setTimeout(15);
-					$rVersions = $rGit->getPreviousVersions($rBaseVersion, 5);
-				} catch (\Throwable $rE) {
-					$rVersions = array();
-				}
-				echo json_encode(array('result' => true, 'current' => $rBaseVersion, 'versions' => $rVersions));
-
-				exit();
-			}
-
-			echo json_encode(array('result' => false));
-
-			exit();
-		}
-		if (RequestManager::get('action') == 'server') {
-			if (Authorization::check('adv', 'edit_server')) {
-				$rSub = RequestManager::get('sub');
-
-				if ($rSub == 'delete') {
-					if (isset($rServers[RequestManager::get('server_id')]) && $rServers[RequestManager::get('server_id')]['is_main'] == 0) {
-						ServerRepository::deleteById(RequestManager::get('server_id'));
-						echo json_encode(array('result' => true));
-
-						exit();
-					}
-
-					echo json_encode(array('result' => false));
-
-					exit();
-				}
-
-				if ($rSub == 'update') {
-					if (!is_numeric(RequestManager::get('server_id'))) {
-						$rIDs = json_decode(RequestManager::get('server_id'), true);
-					} else {
-						$rIDs = array(intval(RequestManager::get('server_id')));
-					}
-
-					foreach ($rIDs as $rID) {
-						$db->query('INSERT INTO `signals`(`server_id`, `time`, `custom_data`) VALUES(?, ?, ?);', $rID, time(), json_encode(array('action' => 'update')));
-					}
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'rollback') {
-					$rVersion = trim((string) RequestManager::get('version'));
-
-					if (!preg_match('/^\d+\.\d+\.\d+$/', $rVersion) || version_compare($rVersion, XC_VM_VERSION, '>=')) {
-						echo json_encode(array('result' => false, 'error' => 'invalid_version'));
-
-						exit();
-					}
-
-					if (!is_numeric(RequestManager::get('server_id'))) {
-						$rIDs = json_decode(RequestManager::get('server_id'), true);
-					} else {
-						$rIDs = array(intval(RequestManager::get('server_id')));
-					}
-
-					foreach ($rIDs as $rID) {
-						$db->query('INSERT INTO `signals`(`server_id`, `time`, `custom_data`) VALUES(?, ?, ?);', $rID, time(), json_encode(array('action' => 'rollback', 'version' => $rVersion)));
-					}
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'enable') {
-					$db->query('UPDATE `servers` SET `enabled` = 1 WHERE `id` = ?;', RequestManager::get('server_id'));
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'disable') {
-					$db->query('UPDATE `servers` SET `enabled` = 0 WHERE `id` = ? AND `is_main` = 0;', RequestManager::get('server_id'));
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'enable_proxy') {
-					$db->query('UPDATE `servers` SET `enable_proxy` = 1 WHERE `id` = ?;', RequestManager::get('server_id'));
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'disable_proxy') {
-					$db->query('UPDATE `servers` SET `enable_proxy` = 0 WHERE `id` = ?;', RequestManager::get('server_id'));
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'kill') {
-					if (SettingsManager::get('redis_handler')) {
-						foreach (ConnectionTracker::getRedisConnections(null, RequestManager::get('server_id'), null, true, false, false) as $rConnection) {
-							ConnectionTracker::closeConnection($rConnection);
-						}
-					} else {
-						$db->query('SELECT * FROM `lines_live` WHERE `server_id` = ?;', RequestManager::get('server_id'));
-
-						foreach ($db->get_rows() as $rRow) {
-							ConnectionTracker::closeConnection($rRow);
-						}
-					}
-
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'restart') {
-					$rStreamIDs = array();
-					$db->query('SELECT `stream_id` FROM `streams_servers` WHERE `server_id` = ? AND `on_demand` = 0 AND `monitor_pid` > 0 AND `pid` > 0 AND `stream_status` = 0;', RequestManager::get('server_id'));
-
-					if (0 >= $db->num_rows()) {
-					} else {
-						foreach ($db->get_rows() as $rRow) {
-							$rStreamIDs[] = intval($rRow['stream_id']);
-						}
-					}
-
-					if (0 >= count($rStreamIDs)) {
-					} else {
-						$rResult = ApiClient::request(array('action' => 'stream', 'sub' => 'start', 'stream_ids' => array_values($rStreamIDs), 'servers' => array(intval(RequestManager::get('server_id')))));
-					}
-
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'start') {
-					$rStreamIDs = array();
-					$db->query('SELECT `stream_id` FROM `streams_servers` WHERE `server_id` = ? AND `on_demand` = 0;', RequestManager::get('server_id'));
-
-					if (0 >= $db->num_rows()) {
-					} else {
-						foreach ($db->get_rows() as $rRow) {
-							$rStreamIDs[] = intval($rRow['stream_id']);
-						}
-					}
-
-					if (0 >= count($rStreamIDs)) {
-					} else {
-						$rResult = ApiClient::request(array('action' => 'stream', 'sub' => 'start', 'stream_ids' => array_values($rStreamIDs), 'servers' => array(intval(RequestManager::get('server_id')))));
-					}
-
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'stop') {
-					$rStreamIDs = array();
-					$db->query('SELECT `stream_id` FROM `streams_servers` WHERE `server_id` = ? AND `on_demand` = 0;', RequestManager::get('server_id'));
-
-					if (0 >= $db->num_rows()) {
-					} else {
-						foreach ($db->get_rows() as $rRow) {
-							$rStreamIDs[] = intval($rRow['stream_id']);
-						}
-					}
-
-					if (0 >= count($rStreamIDs)) {
-					} else {
-						$rResult = ApiClient::request(array('action' => 'stream', 'sub' => 'stop', 'stream_ids' => array_values($rStreamIDs), 'servers' => array(intval(RequestManager::get('server_id')))));
-					}
-
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				echo json_encode(array('result' => false));
-				exit();
-			} else {
-				echo json_encode(array('result' => false));
-				exit();
-			}
-		}
-		if (RequestManager::get('action') == 'proxy') {
-			if (Authorization::check('adv', 'edit_server')) {
-				$rSub = RequestManager::get('sub');
-
-				if ($rSub == 'delete') {
-					ServerRepository::deleteById(RequestManager::get('server_id'));
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'enable') {
-					$db->query('UPDATE `servers` SET `enabled` = 1 WHERE `id` = ?;', RequestManager::get('server_id'));
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'disable') {
-					$db->query('UPDATE `servers` SET `enabled` = 0 WHERE `id` = ?;', RequestManager::get('server_id'));
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'update') {
-					$db->query('INSERT INTO `signals`(`server_id`, `time`, `custom_data`) VALUES(?, ?, ?);', RequestManager::get('server_id'), time(), json_encode(array('action' => 'update')));
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'kill') {
-					if (SettingsManager::get('redis_handler')) {
-						foreach (ServerRepository::getAll()[$rServerID]['parent_id'] as $rParentID) {
-							foreach (ConnectionTracker::getRedisConnections(null, $rParentID, null, true, false, false) as $rConnection) {
-								if ($rConnection['proxy_id'] != RequestManager::get('server_id')) {
-								} else {
-									ConnectionTracker::closeConnection($rConnection);
-								}
-							}
-						}
-					} else {
-						$db->query('SELECT * FROM `lines_live` WHERE `proxy_id` = ?;', RequestManager::get('server_id'));
-
-						foreach ($db->get_rows() as $rRow) {
-							ConnectionTracker::closeConnection($rRow);
-						}
-					}
-
-					echo json_encode(array('result' => true));
-
-					exit();
-				} else {
-					echo json_encode(array('result' => false));
-
-					exit();
-				}
-			} else {
-				echo json_encode(array('result' => false));
-
-				exit();
-			}
 		}
 		if (RequestManager::get('action') == 'package') {
 			if (Authorization::check('adv', 'edit_package')) {
@@ -2131,170 +1850,6 @@ if (isset($_SESSION['hash'])) {
 
 			exit();
 		}
-		if (RequestManager::get('action') == 'fingerprint') {
-			if (Authorization::check('adv', 'fingerprint')) {
-				$rData = json_decode(RequestManager::get('data'), true);
-				$rActiveServers = array();
-
-				foreach ($rServers as $rServer) {
-					if ((360 < time() - $rServer['last_check_ago'] || $rServer['status'] == 2) && $rServer['is_main'] == 0 && $rServer['status'] != 3) {
-						$rServerError = true;
-					} else {
-						$rServerError = false;
-					}
-
-					if ($rServer['status'] != 1 || $rServerError) {
-					} else {
-						$rActiveServers[] = $rServer['id'];
-					}
-				}
-
-				if (!(0 < $rData['id'] && 0 < $rData['font_size'] && 0 < strlen($rData['font_color']) && 0 < strlen($rData['xy_offset']) && (0 < strlen($rData['message']) || $rData['type'] < 3))) {
-				} else {
-					if (SettingsManager::get('redis_handler')) {
-						if (isset($rData['user'])) {
-							$rRows = ConnectionTracker::getRedisConnections($rData['id'], null, null, true, false, false);
-						} else {
-							$rRows = ConnectionTracker::getRedisConnections(null, null, $rData['id'], true, false, false);
-						}
-
-						$rUserMap = $rUserIDs = array();
-
-						foreach ($rRows as $rRow) {
-							if (in_array($rRow['user_id'], $rUserIDs)) {
-							} else {
-								$rUserIDs[] = intval($rRow['user_id']);
-							}
-						}
-
-						if (0 >= count($rUserIDs)) {
-						} else {
-							$db->query('SELECT `id`, `username` FROM `lines` WHERE `id` IN (' . implode(',', $rUserIDs) . ');');
-
-							foreach ($db->get_rows() as $rRow) {
-								$rUserMap[$rRow['id']] = $rRow['username'];
-							}
-						}
-					} else {
-						if (isset($rData['user'])) {
-							$db->query('SELECT `lines_live`.`activity_id`, `lines_live`.`uuid`, `lines_live`.`user_id`, `lines_live`.`server_id`, `lines`.`username` FROM `lines_live` LEFT JOIN `lines` ON `lines`.`id` = `lines_live`.`user_id` WHERE `user_id` = ?;', $rData['id']);
-						} else {
-							$db->query('SELECT `lines_live`.`activity_id`, `lines_live`.`uuid`, `lines_live`.`user_id`, `lines_live`.`server_id`, `lines`.`username` FROM `lines_live` LEFT JOIN `lines` ON `lines`.`id` = `lines_live`.`user_id` WHERE `stream_id` = ?;', $rData['id']);
-						}
-
-						$rRows = $db->get_rows();
-					}
-
-					if (count($rRows) > 0) {
-						set_time_limit(360);
-						ini_set('max_execution_time', 360);
-						ini_set('default_socket_timeout', 15);
-
-						foreach ($rRows as $rRow) {
-							if (in_array($rRow['server_id'], $rActiveServers)) {
-								$rArray = array('font_size' => $rData['font_size'], 'font_color' => $rData['font_color'], 'xy_offset' => $rData['xy_offset'], 'message' => '', 'uuid' => $rRow['uuid']);
-
-								if ($rData['type'] == 1) {
-									$rArray['message'] = $rRow['uuid'];
-								} elseif ($rData['type'] == 2) {
-									$rArray['message'] = (SettingsManager::get('redis_handler') ? $rUserMap[$rRow['user_id']] : $rRow['username']);
-								} elseif ($rData['type'] == 3) {
-									$rArray['message'] = $rData['message'];
-								}
-
-								$rArray['action'] = 'signal_send';
-								$rSuccess = ApiClient::systemRequest(intval($rRow['server_id']), $rArray);
-							}
-						}
-					}
-				}
-
-				echo json_encode(array('result' => true));
-
-				exit();
-			} else {
-				echo json_encode(array('result' => false));
-
-				exit();
-			}
-		}
-		if (RequestManager::get('action') == 'restart_all_services') {
-			if (Authorization::check('adv', 'servers')) {
-				foreach ($rServers as $rServer) {
-					if (!$rServer['server_online']) {
-					} else {
-						$db->query("INSERT INTO `signals`(`server_id`, `custom_data`, `time`) VALUES(?, '{\"action\": \"restart_services\"}', ?);", $rServer['id'], time());
-					}
-				}
-				echo json_encode(array('result' => true));
-
-				exit();
-			} else {
-				echo json_encode(array('result' => false));
-
-				exit();
-			}
-		}
-		if (RequestManager::get('action') == 'restart_services') {
-			if (Authorization::check('adv', 'edit_server')) {
-				if (!is_numeric(RequestManager::get('server_id'))) {
-					$rIDs = json_decode(RequestManager::get('server_id'), true);
-				} else {
-					$rIDs = array(intval(RequestManager::get('server_id')));
-				}
-
-				foreach ($rIDs as $rID) {
-					$db->query("INSERT INTO `signals`(`server_id`, `custom_data`, `time`) VALUES(?, '{\"action\": \"restart_services\"}', ?);", $rID, time());
-				}
-				echo json_encode(array('result' => true));
-
-				exit();
-			} else {
-				echo json_encode(array('result' => false));
-
-				exit();
-			}
-		}
-		if (RequestManager::get('action') == 'reboot_server') {
-			if (Authorization::check('adv', 'edit_server')) {
-				if (!is_numeric(RequestManager::get('server_id'))) {
-					$rIDs = json_decode(RequestManager::get('server_id'), true);
-				} else {
-					$rIDs = array(intval(RequestManager::get('server_id')));
-				}
-
-				foreach ($rIDs as $rID) {
-					$db->query("INSERT INTO `signals`(`server_id`, `custom_data`, `time`) VALUES(?, '{\"action\": \"reboot\"}', ?);", $rID, time());
-				}
-				echo json_encode(array('result' => true));
-
-				exit();
-			} else {
-				echo json_encode(array('result' => false));
-
-				exit();
-			}
-		}
-		if (RequestManager::get('action') == 'update_binaries') {
-			if (Authorization::check('adv', 'edit_server')) {
-				if (!is_numeric(RequestManager::get('server_id'))) {
-					$rIDs = json_decode(RequestManager::get('server_id'), true);
-				} else {
-					$rIDs = array(intval(RequestManager::get('server_id')));
-				}
-
-				foreach ($rIDs as $rID) {
-					$db->query("INSERT INTO `signals`(`server_id`, `custom_data`, `time`) VALUES(?, '{\"action\": \"update_binaries\"}', ?);", $rID, time());
-				}
-				echo json_encode(array('result' => true));
-
-				exit();
-			} else {
-				echo json_encode(array('result' => false));
-
-				exit();
-			}
-		}
 		if (RequestManager::get('action') == 'probe_stream') {
 			if (Authorization::check('adv', 'add_stream') || Authorization::check('adv', 'edit_stream')) {
 				$rAnalyseDuration = abs(intval(SettingsManager::get('stream_max_analyze')));
@@ -2788,187 +2343,6 @@ if (isset($_SESSION['hash'])) {
 
 					exit();
 				}
-			} else {
-				echo json_encode(array('result' => false));
-
-				exit();
-			}
-		}
-		if (RequestManager::get('action') == 'server_view') {
-			if (Authorization::check('adv', 'add_server') || Authorization::check('adv', 'edit_server')) {
-				if (isset($rServers[RequestManager::get('server_id')])) {
-					$rServer = $rServers[RequestManager::get('server_id')];
-				} else {
-					if (isset($rProxyServers[RequestManager::get('server_id')])) {
-						$rServer = $rProxyServers[RequestManager::get('server_id')];
-					} else {
-						echo json_encode(array('result' => false));
-
-						exit();
-					}
-				}
-
-				$rStats = array('open_connections' => 0, 'total_running_streams' => 0, 'online_users' => 0, 'offline_streams' => 0, 'gpu_info' => json_decode($rServer['gpu_info'], true), 'watchdog' => json_decode($rServer['watchdog_data'], true));
-				$rStats['open_connections'] = ($rServer['connections'] ?: 0);
-				$rStats['online_users'] = ($rServer['users'] ?: 0);
-				$db->query('SELECT COUNT(*) AS `count` FROM `streams_servers` LEFT JOIN `streams` ON `streams`.`id` = `streams_servers`.`stream_id` WHERE `server_id` = ? AND `pid` > 0 AND `type` = 1;', $rServer['id']);
-				$rStats['total_running_streams'] = $db->get_row()['count'];
-				$db->query('SELECT COUNT(*) AS `count` FROM `streams_servers` LEFT JOIN `streams` ON `streams`.`id` = `streams_servers`.`stream_id` WHERE `server_id` = ? AND `type` = 1 AND ((`streams_servers`.`monitor_pid` IS NOT NULL AND `streams_servers`.`monitor_pid` > 0) AND (`streams_servers`.`pid` IS NULL OR `streams_servers`.`pid` <= 0) AND `streams_servers`.`stream_status` <> 0);', $rServer['id']);
-				$rStats['offline_streams'] = $db->get_row()['count'];
-				echo json_encode(array('result' => true, 'data' => $rStats, 'netspeed' => (intval($rServer['network_guaranteed_speed']) ?: 1000)));
-
-				exit();
-			}
-
-			echo json_encode(array('result' => false));
-
-			exit();
-		}
-		if (RequestManager::get('action') == 'server_stats') {
-			if (Authorization::check('adv', 'add_server') || Authorization::check('adv', 'edit_server')) {
-				$rID = intval(RequestManager::get('id'));
-
-				if (isset($rServers[$rID])) {
-					$rWatchdog = ProcessChecker::getWatchdog($rID);
-					$rReturn = array();
-
-					foreach ($rWatchdog as $rData) {
-						$rReturn[] = array('cpu' => $rData['cpu'], 'memory' => $rData['total_mem_used_percent'], 'input' => $rData['bytes_received'], 'output' => $rData['bytes_sent'], 'date' => $rData['time']);
-					}
-					echo json_encode(array('result' => true, 'data' => $rReturn));
-
-					exit();
-				} else {
-					echo json_encode(array('result' => false));
-
-					exit();
-				}
-			} else {
-				echo json_encode(array('result' => false));
-
-				exit();
-			}
-		}
-		if (RequestManager::get('action') == 'rtmp_kill') {
-			if (Authorization::check('adv', 'rtmp')) {
-				echo ApiClient::systemRequest(intval(RequestManager::get('server')), array('action' => 'rtmp_kill', 'name' => RequestManager::get('name')));
-
-				exit();
-			}
-
-			echo json_encode(array('result' => false));
-
-			exit();
-		}
-		if (RequestManager::get('action') == 'install_status') {
-			if (Authorization::check('adv', 'add_server') || Authorization::check('adv', 'edit_server')) {
-				$rServers = ServerRepository::getAll(true);
-				$rServerID = intval(RequestManager::get('server_id'));
-				$rFilename = BIN_PATH . 'install/' . $rServerID . '.install';
-
-				if (file_exists($rFilename)) {
-					echo json_encode(array('result' => true, 'data' => trim(file_get_contents($rFilename)), 'status' => intval($rServers[$rServerID]['status'])));
-
-					exit();
-				}
-
-				echo json_encode(array('result' => false));
-
-				exit();
-			}
-
-			echo json_encode(array('result' => false));
-
-			exit();
-		}
-		if (RequestManager::get('action') == 'reinstall_server') {
-			if (Authorization::check('adv', 'add_server') || Authorization::check('adv', 'edit_server')) {
-				$rServerID = intval(RequestManager::get('server_id'));
-
-				if ($rServers[$rServerID]['server_type'] == 0) {
-					$rType = 2;
-				} else {
-					$rType = 1;
-				}
-
-				$rFilename = BIN_PATH . 'install/' . $rServerID . '.json';
-
-				if (file_exists($rFilename)) {
-					$rParams = json_decode(file_get_contents($rFilename), true);
-					$db->query('UPDATE `servers` SET `status` = 3 WHERE `id` = ?;', $rServerID);
-
-					if (isset($rParams['http_broadcast_port'])) {
-						$rCommand = PHP_BIN . ' ' . MAIN_HOME . 'console.php server:install ' . $rType . ' ' . intval($rServerID) . ' ' . intval($rParams['ssh_port']) . ' ' . escapeshellarg($rParams['root_username']) . ' ' . escapeshellarg($rParams['root_password']) . ' ' . intval($rParams['http_broadcast_port']) . ' ' . intval($rParams['https_broadcast_port']) . ' > "' . BIN_PATH . 'install/' . intval($rServerID) . '.install" 2>/dev/null &';
-					} else {
-						$rCommand = PHP_BIN . ' ' . MAIN_HOME . 'console.php server:install ' . $rType . ' ' . intval($rServerID) . ' ' . intval($rParams['ssh_port']) . ' ' . escapeshellarg($rParams['root_username']) . ' ' . escapeshellarg($rParams['root_password']) . ' > "' . BIN_PATH . 'install/' . intval($rServerID) . '.install" 2>/dev/null &';
-					}
-
-					shell_exec($rCommand);
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				echo json_encode(array('result' => false));
-
-				exit();
-			}
-
-			echo json_encode(array('result' => false));
-
-			exit();
-		}
-		if (RequestManager::get('action') == 'fpm_status') {
-			if (Authorization::check('adv', 'add_server') || Authorization::check('adv', 'edit_server')) {
-				$rData = str_replace("\n", '<br/>', ApiClient::systemRequest(RequestManager::get('server_id'), array('action' => 'fpm_status')));
-
-				if (empty($rData)) {
-					$rData = '<strong>No response from status page.</strong>';
-				} else {
-					$rInstances = intval($rServers[RequestManager::get('server_id')]['total_services']);
-
-					if (!$rInstances) {
-					} else {
-						$rData .= '<br/><br/><strong>Results from 1 of ' . $rInstances . ' PHP-FPM instances</strong>';
-					}
-				}
-
-				echo json_encode(array('result' => true, 'data' => $rData));
-
-				exit();
-			}
-
-			echo json_encode(array('result' => false));
-
-			exit();
-		}
-		if (RequestManager::get('action') == 'update_all_servers') {
-			if (Authorization::check('adv', 'servers')) {
-				foreach ($rServers as $rServer) {
-					if ($rServer['server_online']) {
-						$db->query('INSERT INTO `signals`(`server_id`, `time`, `custom_data`) VALUES(?, ?, ?);', $rServer['id'], time(), json_encode(array('action' => 'update')));
-					}
-				}
-				echo json_encode(array('result' => true));
-
-				exit();
-			} else {
-				echo json_encode(array('result' => false));
-
-				exit();
-			}
-		}
-		if (RequestManager::get('action') == 'update_all_binaries') {
-			if (Authorization::check('adv', 'servers')) {
-				foreach ($rServers as $rServer) {
-					if (!$rServer['server_online']) {
-					} else {
-						$db->query('INSERT INTO `signals`(`server_id`, `time`, `custom_data`) VALUES(?, ?, ?);', $rServer['id'], time(), json_encode(array('action' => 'update_binaries')));
-					}
-				}
-				echo json_encode(array('result' => true));
-
-				exit();
 			} else {
 				echo json_encode(array('result' => false));
